test(generic-scoring): split reconciler test by scenario
The reconciler test had grown to 509 lines / 65 expects in a single file,
mixing pure-logic predicates, DDB transition scenarios, pagination cursors,
stuck-DELETING rescue, and ConditionalCheckFailedException skip behavior.

Split per #1255 into 5 scenario-focused files (all <200 lines), each with a
single clear focus. Shared `buildCtx()` / `NOW_ISO` go into a sibling
`.test-helpers.ts` (not `.test.ts`, so vitest does not collect it).

Closes #1255
